Enhance lsp attach
This is primarily to fix the issue where LsapAttach was running multiple
times for the same buffer, once for each LSP. In the case of csharp, it
was omnisharp and copilot. This meant that keymaps were being set
multiple times, sometimes overwriting the omnisharp-specific ones.

This commit changes it to only add keybinds for an lsp client if the
client supports that capability.

It also comments out a few unused c# LSP things, like csharp_ls and the
new roslyn LSP.